+++ CHANGES.rst
@@ -31,6 +31,7 @@ Improvements
 - Log user merges to user log (:issue:`6882`, :pr:`6920`)
 - Allow re-sending emails from their log entries (:issue:`6805`, :pr:`6909`,
   thanks :user:`duartegalvao, unconventionaldotdev`)
+- Add a text overflow option to the badge designer (:pr:`6944`, thanks :user:`SegiNyn`)
 
 Bugfixes
 ^^^^^^^^